Some fixes for running dialog flows on top of other applications calling app manager with mime_open calls.
* src/util.cc, src/util.h (end_dialog_flow, mark_keep_running):
added functions to quit the application when a dialog flow ends
and nobody requested to keep the application running. Added also
a function to request to keep running.

* src/main.cc (window_property_topmost_notify_event, main): mark 
the application to keep running after a dialog flow when the main 
window becomes topmost.

Other fixes:

* src/main.cc (install_from_file_reply): removed call to
end_dialog_flow in this function, that made application die before
it should.


git-svn-id: https://garage.maemo.org/svn/hildon-app-mgr/trunk@94 e55152ac-3a23-0410-980c-f88739b87167