Stop hildon-status-menu after setting the device in offline mode.
* src/dbus.cc (set_device_mode): Use a blocking dbus message for
setting the device mode as we must be sure the message was properly
handled before continuing with anything else.
* src/operations.cc (ip_stop_hsm_and_install_delayed): New GSource
func to stop hildon-status-menu and continue installing the package.
(ip_install_cur): For SSU packages, call the new function
above after some small delay (3 secs), to avoid problems with previous
statement to set device in offline mode. If not a SSU package, just
continue installing it as usual.