Bear in mind result of rescue mode to show SSU banner on boot
* src/apt-worker.cc (RESCUE_RESULT_FILE): New path for the file
storing the result (1 or 0) of the last rescue operation.
(do_rescue): Always write the content of RESCUE_RESULT_FILE either
with 1 or 0 to reflect status after rescuing (success or not).

* src/ham-after-boot.c (main): Apart from checking the UFILE_BOOT
file, also check RESCUE_RESULT_FILE content to actually know when it
really makes sense to show the "SSU successfully" banner or not